+++ jsconfig.json
@@ -1,0 +1,35 @@
+{
+  "compilerOptions": {
+    "checkJs": false,
+    "target": "ES2022",
+    "module": "ES2022",
+    "moduleResolution": "Bundler",
+    "baseUrl": ".",
+    "paths": {
+      "@h2/*": ["app/components/hydrogen/*"],
+      "~/*": ["app/*"]
+    }
+  },
+  // "compilerOptions": {
+  //   "lib": ["DOM", "DOM.Iterable", "ES2022"],
+  //   "isolatedModules": true,
+  //   "esModuleInterop": true,
+  //   "jsx": "react-jsx",
+  //   "moduleResolution": "Bundler",
+  //   "resolveJsonModule": true,
+  //   "module": "ES2022",
+  //   "target": "ES2022",
+  //   "strict": true,
+  //   "allowJs": true,
+  //   "forceConsistentCasingInFileNames": true,
+  //   "skipLibCheck": true,
+  //   "baseUrl": ".",
+  //   "types": ["@shopify/oxygen-workers-types"],
+  //   "paths": {
+  //     "@h2/*": ["app/components/hydrogen/*"],
+  //     "~/*": ["app/*"]
+  //   },
+  //   "noEmit": true
+  // },
+  "include": ["./**/*.d.ts", "./**/*.js", "./**/*.jsx"]
+}
